leetcode-sharp/2300/S02319.cs
namespace leetcode_sharp;

// 2319. Check if Matrix Is X-Matrix
// https://leetcode.com/problems/check-if-matrix-is-x-matrix/
public class S02319
{
    public bool CheckXMatrix(int[][] grid)
    {
        var n = grid.Length;

        for (var i = 0; i < n; i++)
        {
            for (var j = 0; j < n; j++)
            {
                if (i == j || i == n - j - 1)
                {
                    if (grid[i][j] == 0)
                    {
                        return false;
                    }
                }
                else
                {
                    if (grid[i][j] != 0)
                    {
                        return false;
                    }
                }
            }
        }

        return true;
    }
}